Fixed test values

+++ src/test/java/com/bernardomg/tabletop/dreadball/web/toolkit/test/integration/codex/service/ITDefaultSponsorBuilderServiceValidateTeam.java
@@ -161,7 +161,7 @@ public final class ITDefaultSponsorBuilderServiceValidateTeam
 
         result = service.validateTeam(selection);
 
-        Assert.assertEquals(new Integer(17), result.getValoration());
+        Assert.assertEquals(new Integer(16), result.getValoration());
     }
 
     /**
